Moved library functions from `PageGeneratorPlugin.py` to `render_utils.py` and `date_utils.py`
Also did some other refactoring, e.g. moving `soft_escape` from `game.py`.
